Added: new ignore lines
These default paths are ones generated by the mains script of this project if
run within cloned directory. While this is not the suggested actions for
running the script the authors are at least attemting to prevent users from
accedentally pushing any private data, no matter how encrypted it may be.