More atempts at fixes
Yet again lots of changes to build script stuff for Travis-CI and Code
Climate. Not much otherwise to comment on this commit. As of latest main
script version "date -u +%s" most of the features asside from color output
should be working for most Bash shells that have mapfile and gnupg
available.