More Travis-CI attempts at fixes
Some of the readers of these commit comments may wonder at why the authors
are so push happy. Well for one reason the build scripts should provide
sufficent guidence for advanced users of this project to get up and running
within a test enviroment. And the second reason is that both power and
network access where disconected within the same week, thus as the authors
of this project may not be able to continue maintaining this project with
the previous push reagularoty and have focused on making this project work
for readers before these two naccesityies (first world problems) are
compleatly cut off from the original authors... enjoy while the project is
still available and maintained actively.