Added to documentation
Now includes instructions and examples on how to set variables or command
line values for main script. This reflects the edits made to the main
script, within the case statment that handles arguemtns, to allow setting
either reliably when reconized options or variables are used... The authors
have not tested if these can be mixed and matched on the same command line
operation but based on how the loops process input this should be
compatible with mixing it up.