Re-disabled one line within main script
Based off tests it appieres as though setting arguemtns as an array and
passing that array as an arguemnt is interducing an extra new line at the
end. This is causing isssues with testing script copy processes because the
main script tries to write null data to a non-listening pipe file.