Minor fixes to main script
Eventually the authors will be making additions to control the permissions
of encrypted files seperate from permissions of self generated logs made by
this script, until then, log rotation actions that re-build the encrypted
output file have now been added prior to writing to the output file... in
other words if the output does not exsist then an empty file with proper
permissions is made prior to trying to output to it. These addtions may
allow for full auto-builds to now succeed. In related news; the Travis-CI
team has gotten back to the authors of this project (the oddities of local
build vs remote builds warented making premtive contact) and have been
wonderful work with.