Edited badges on main ReadMe
The new addition was made via the helper script for formatting the links
corectly, should work just fine for being a link to the main script.
However, updating the ReadMe.md file is a manual process that is done after
updating the version in the main script... So another helper script maybe
written in order to make editing both files verions a bit more streamlined.