Added new helper script
This new script combines GnuPG notes found within the documentation
directory of this project for users that either wish to contribute back to
this project or make use of the customizations used by the authors of this
project. Eventually this script will replace protions of the Travic-CI
build scripts that preform similar functions.