Getting closer...
The last log, for Commit 391b125, got really close to generating a revoke
certificate. Looks as though the authors have placed the passphrase where
GnuPG expected it but by prefixing the option for passphrase, like that for
generating the keys in the first place, seems to be causing read errors.
Fingers crossed that this push does it.