Mmm, that last build was good...
Alrighty, updated status of helper script in related readme file, removed
no longer needed lines from encryption build script, and hopefully
corrected for the oddness reported by 'diff' command in decryption build
script. Where the authors still capible of dancing they would be, readers
that are able to are encuraged to at this time in celobration of a fully
tested and working project.