Fixing unset variable/array errors
On close inspection it appiers as though the first line is correctly being
decrypted, further more the helper script appiers to be correctly pushing
the 'n+1' block of encrypted data correctly to the named pipe. However, it
looks as though this is where breakage is happening as the sencond+n blocks
are not being read.