So about that celobration?...
The last remote builds showed that the helper script is now fully
functional and that GnuPG file descriptor issues can now be resolved. This
is now a test of new command line options that where added to helper script
in order to make it conform with scripting style already present within
this project.

Side note here and a big thank you to the Travis-CI teem. Not only is the
automation of build process very much apprechiated but also the swiftness
in responce time to emailed corispondence. This totality in aiming to aid
developers earns the Travis-CI team very high marks/grades; in this
projects authors oppinion that is.