Added the begining of un-padding function
Currently only 'append' and 'prepend' options are written for testing,
still working on 'above' and 'bellow' options that the main script of this
project can use for adding junk to read input. While the rest of this
script is in OK condition this one function is going to be funky.