Many changes to decryption build/helper scripts
Moved logics for handling bulk input/output directories to helper script as
well as made code a bit more easy to maintain in case of errors or new
features to be added. See lines around internal scripte comments titled
'TO-DO' for authors notes of what we are still working on. This build may
fail, if so then the build script maybe rolled-back to previous version for
comperison localy.