Changed build script copy encryption variables
Modified script copy variables to be more in-line with main script
operations. Removed premature exits unless script copy is not executable,
if file or not a file then exits still happen but this time prior to any
attempts at writes (this also was edited into the main script) such that
non-executable versions can be saved for inspection automaticly.